+++ src/schema-generator/index.ts
@@ -1,2 +1,5 @@
-export { jsonToSchema, addRule } from "./json-to-schema";
+import { YupHelper } from "./yup-helper";
+export { jsonToSchema } from "./json-to-schema";
 export * from "./types";
+
+export const addRule = YupHelper.addRule;
